+++ brownie/tests/datastructures/__init__.py
@@ -12,7 +12,7 @@ from __future__ import with_statement
 
 from attest import Tests, TestBase, test, Assert
 
-from brownie.datastructures import missing
+from brownie.datastructures import missing, StackedObject
 from brownie.tests.datastructures import (sets, queues, sequences, mappings,
                                           iterators)
 
@@ -28,7 +28,30 @@ class TestMissing(TestBase):
         Assert(repr(missing)) == 'missing'
 
 
+class TestStackedObject(TestBase):
+    @test
+    def stacking(self):
+        s = StackedObject([])
+        with Assert.raises(AttributeError):
+            s.foo
+
+        with Assert.raises(RuntimeError):
+            s.pop()
+
+        s.push({'foo': False})
+        Assert(s.foo) == False
+        s.push({'foo': True})
+        Assert(s.foo) == True
+        s.pop()
+        Assert(s.foo) == False
+
+    @test
+    def repr(self):
+        s = StackedObject([{}])
+        Assert(repr(s)) == 'StackedObject([{}])'
+
+
 tests = Tests([
-    TestMissing, queues.tests, sets.tests, sequences.tests, mappings.tests,
-    iterators.tests
+    TestMissing, TestStackedObject, queues.tests, sets.tests, sequences.tests,
+    mappings.tests, iterators.tests
 ])